chore: Use relative state include
My use case is that I actually have to rename this formula because the
SUSE Manager ships with its own `packages` formula that it uses
internally. But I still want to use this formula in combination. So one
of them has to be renamed. This change makes that way easier.

=== Do not change lines below ===
{
 "chain": [],
 "cmd": "git ls-files -z packages | xargs --null -I '()' find '()' -type f -print0 | xargs --null sed --in-place --regexp-extended 's#packages/#./#g;s#- packages\\.#- .#;'",
 "exit": 0,
 "extra_inputs": [],
 "inputs": [],
 "outputs": [],
 "pwd": "."
}
^^^ Do not change lines above ^^^